Use function wrapper replacements by default.
