+++ src/fs_util.ts
@@ -1,0 +1,27 @@
+import * as fs from 'fs';
+import * as path from 'path';
+
+export function readFilesSync(dir: string): {[key: string]: string} | null {
+  let foundAnything = false
+  const files: {[key: string]: string} = {};
+
+  fs.readdirSync(dir).forEach(filename => {
+    const name = path.parse(filename).name;
+    const filepath = path.resolve(dir, filename);
+    const stat = fs.statSync(filepath);
+    const isFile = stat.isFile();
+
+    if (isFile) {
+         const content = fs.readFileSync(path.join(dir, filename))
+         files[name] = content.toString()
+         foundAnything = true
+    }
+  });
+
+  if (foundAnything) {
+    return files;
+  } else {
+    return null
+  }
+
+}
